# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/institutions": {
      "get": {
        "operationId": "InstitutionsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 25,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Serbest metin arama",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "gender",
            "required": false,
            "in": "query",
            "schema": {
              "enum": ["FEMALE", "MALE"],
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["institutions"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "InstitutionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInstitutionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["institutions"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/institutions/{id}": {
      "get": {
        "operationId": "InstitutionsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["institutions"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "InstitutionsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInstitutionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["institutions"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "InstitutionsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": ["institutions"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/groups": {
      "get": {
        "operationId": "GroupsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 25,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Serbest metin arama",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "institutionId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "termId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scholarshipProgramId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["groups"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "GroupsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["groups"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/groups/{id}": {
      "get": {
        "operationId": "GroupsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["groups"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "GroupsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["groups"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "GroupsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": ["groups"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/groups/{id}/members": {
      "get": {
        "operationId": "GroupsController_members",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": false,
            "in": "query",
            "description": "Bu gunde grupta olanlar; bos ise bugun",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["groups"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "GroupsController_addMembers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMembersDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["groups"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/groups/{id}/members/{studentId}": {
      "delete": {
        "operationId": "GroupsController_endMembership",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "studentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "effectiveTo",
            "required": false,
            "in": "query",
            "description": "Gruptaki son gun + 1 (bos ise bugun)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": ["groups"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/holidays": {
      "get": {
        "operationId": "HolidaysController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 25,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Serbest metin arama",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "institutionId",
            "required": false,
            "in": "query",
            "description": "Bu yurda ait + tum yurtlara gecerli tatiller",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["holidays"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "HolidaysController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHolidayDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["holidays"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/holidays/{id}": {
      "delete": {
        "operationId": "HolidaysController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": ["holidays"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Öğrenci Takip Sistemi API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateInstitutionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "TDV Erkek Öğrenci Yurdu Ankara"
          },
          "code": {
            "type": "string",
            "example": "ANK-ERKEK-1",
            "description": "Buyuk harf, rakam ve tire"
          },
          "gender": {
            "type": "string",
            "enum": ["FEMALE", "MALE"],
            "description": "Yurttaki ogrencilerin cinsiyeti"
          }
        },
        "required": ["name", "code"]
      },
      "UpdateInstitutionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "TDV Erkek Öğrenci Yurdu Ankara"
          },
          "code": {
            "type": "string",
            "example": "ANK-ERKEK-1",
            "description": "Buyuk harf, rakam ve tire"
          },
          "gender": {
            "type": "string",
            "enum": ["FEMALE", "MALE"],
            "description": "Yurttaki ogrencilerin cinsiyeti"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
      "CreateGroupDto": {
        "type": "object",
        "properties": {
          "institutionId": {
            "type": "string",
            "description": "Sadece sistem yoneticisi icin zorunlu"
          },
          "termId": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "İlahiyat Akademi - A Grubu"
          },
          "scholarshipProgramId": {
            "type": "string",
            "description": "Bos ise grupta farkli programlardan ogrenci olabilir"
          }
        },
        "required": ["termId", "name"]
      },
      "UpdateGroupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "termId": {
            "type": "string",
            "description": "Uyelik/ders gecmisi olan grupta degistirilemez"
          },
          "scholarshipProgramId": {
            "type": "object",
            "nullable": true,
            "description": "Uyelik/ders gecmisi olan grupta degistirilemez"
          }
        }
      },
      "AddMembersDto": {
        "type": "object",
        "properties": {
          "studentIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "effectiveFrom": {
            "type": "string",
            "example": "2026-10-01"
          }
        },
        "required": ["studentIds", "effectiveFrom"]
      },
...
      "CreateHolidayDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2026-10-29"
          },
          "description": {
            "type": "string",
            "example": "Cumhuriyet Bayramı"
          },
          "allInstitutions": {
            "type": "boolean",
            "description": "true: tum yurtlar (resmi tatil, sadece sistem yoneticisi)"
          },
          "institutionId": {
            "type": "string",
            "description": "Yurda ozel tatil; sistem yoneticisi icin zorunlu"
          }
        },
        "required": ["date", "description"]
      },
...
    }
  }
}